refactor: reorganize package structure for GUI screens
