+++ src/main/java/edu/ntnu/iir/bidata/view/gui/screens/ChooseBoardScreen.java
@@ -1,8 +1,11 @@
-package edu.ntnu.iir.bidata.view.gui;
+package edu.ntnu.iir.bidata.view.gui.screens;
 
 import edu.ntnu.iir.bidata.model.Board;
 import edu.ntnu.iir.bidata.model.GameType;
 import edu.ntnu.iir.bidata.view.AppEvent;
+import edu.ntnu.iir.bidata.view.gui.BoardCanvas;
+import edu.ntnu.iir.bidata.view.gui.BoardCanvasFactory;
+import edu.ntnu.iir.bidata.view.gui.GUIApp;
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
