+++ src/main/java/edu/ntnu/iir/bidata/view/gui/screens/HomeScreen.java
@@ -1,5 +1,6 @@
-package edu.ntnu.iir.bidata.view.gui;
+package edu.ntnu.iir.bidata.view.gui.screens;
 
+import edu.ntnu.iir.bidata.view.gui.GUIApp;
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
